packages/agents/pyproject.toml
[project]
name = "swarm-agents"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-ollama>=0.2.0",
    "web3>=7.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]
